requirements.txt
streamlit
langchain>=0.2.0
langchain-community
faiss-cpu
PyPDF2
python-dotenv
huggingface_hub==0.17.1
requests
sentence-transformers>=2.7.0